Basic api server (#4)
* Created the API Routes

This is a unclean version. But all of the endpoints are reachable and the correct methods.
In the end I will want to split things into modules and add handler functions.
An ok first draft.

* Put things in modules

* Created api module

This module contians pretty much everything. It has a few more nested modules within it.

* Better api returns

Now using warp::reply::with_status. This allows a response to contrain a status code and
message. There is also a reply fucntion with will send json, this will be needed later.